:art: Update: 구조 변경, 고객 대쉬보드 형식
